Refactor folder structure and move  components to protectedAudience
