+++ packages/explorable-explanations/src/index.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './components';
+export * from './protectedAudience/components';
 export {
   app,
   userSketch,
