+++ packages/explorable-explanations/src/protectedAudience/components/rippleEffect.js
@@ -16,8 +16,8 @@
 /**
  * Internal dependencies.
  */
-import app from '../protectedAudience/app';
-import config from '../protectedAudience/config';
+import app from '../app';
+import config from '../config';
 
 const RippleEffect = async ({ x, y }) => {
   rippleEffect.setUp();
